Reduce thread creation and runtime
Instead of creating a new thread for each new test, create as many
threads as the computer can handle and continue using those threads.
This removes the need for CountDownLatch, ExecutorService, Executors.

Make time a BigInteger instead of a long in case it is unable to handle
larger test cases.

Replace Math.random() with ThreadLocalRandom to avoid possible
contention.